Move VersionedHash to a better place
